Rename checks_newstyle/ to checks_unlisted/
